# ONGConnect/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ONGConnect — Match Voluntários e ONGs",
        "description": "API REST para conexão entre ONGs do Alto Vale do Itajaí e voluntários da comunidade. Projeto de Extensão — Unidavi 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/inscricoes/minhas": {
            "get": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Lista as inscrições do voluntário autenticado",
                "operationId": "8258352c36650297b788595588652f0c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de inscrições"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Apenas voluntários têm inscrições"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/demandas/{id}/inscricoes": {
            "get": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Lista as inscrições de uma demanda (apenas a ONG dona)",
                "operationId": "dc88610e22d0b3e0a41ac0de41673a03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de inscrições"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Demanda não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Inscreve o voluntário autenticado em uma demanda",
                "operationId": "14c1f1782dcc8b8037a530e0fd2922ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mensagem": {
                                        "type": "string",
                                        "example": "Tenho disponibilidade nos fins de semana."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscrição realizada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Apenas voluntários podem se inscrever"
                    },
                    "404": {
                        "description": "Demanda não encontrada"
                    },
                    "422": {
                        "description": "Demanda fechada, sem vagas ou já inscrito"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inscricoes/{id}/aceitar": {
            "post": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Aceita uma inscrição pendente (ONG dona da demanda)",
                "operationId": "3bd7ac9598425600e5bdae1f9dc43161",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscrição aceita"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Inscrição não encontrada"
                    },
                    "422": {
                        "description": "Inscrição não está pendente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inscricoes/{id}/recusar": {
            "post": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Recusa uma inscrição pendente (ONG dona da demanda)",
                "operationId": "01713093238a812f98c824d2e85f5387",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscrição recusada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Inscrição não encontrada"
                    },
                    "422": {
                        "description": "Inscrição não está pendente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inscricoes/{id}/concluir": {
            "post": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Conclui uma inscrição aceita — libera avaliação (ONG dona)",
                "operationId": "578e71aae706f0a3a4e99062da9c5f89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscrição concluída — avaliação liberada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Inscrição não encontrada"
                    },
                    "422": {
                        "description": "Inscrição não está aceita"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inscricoes/{id}/cancelar": {
            "post": {
                "tags": [
                    "Inscricoes"
                ],
                "summary": "Cancela a própria inscrição (voluntário — se pendente ou aceita)",
                "operationId": "ac5e8b0ad666c6f6a7464055439ca4f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscrição cancelada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Inscrição não encontrada"
                    },
                    "422": {
                        "description": "Inscrição não pode ser cancelada neste status"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticação via token Bearer (Laravel Sanctum)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}